discourse: 3.3.0.beta2
discourse/plugins: init multiple plugins, update existing, remove zendesk plugin

added/updated automatically using update.py

New standard plugins:
* discourse-adplugin
* discourse-akismet
* discourse-apple-auth
* discourse-cakeday
* discourse-footnote
* discourse-patreon
* discourse-subscriptions

New business plugins:
* discourse-linkedin-auth
* discourse-lti
* discourse-automation
* discourse-policy
* discourse-gamification

New enterprise plugins:
* discourse-graphviz
* discourse-tooltips
* discourse-saml

(cherry picked from commit da8a1470fba77708cd6a025725dea0e41727d21a)

discourse/plugins: init discourse-events

https://meta.discourse.org/t/events-plugin/69776
(cherry picked from commit 661574c2c68cb4b62e66978507b0e9bf5c683d7f)

discourse/plugins: init discourse-follow

(cherry picked from commit 0a277dc044755d04b19721e89d4a6213f1c60d79)

discourse/plugins: init discourse-push-notifications

(cherry picked from commit 2bee700c619122d4e0f3c86d5fbda249fdd36bf1)

discourse/plugins: init discourse-whos-online

(cherry picked from commit 15cf10d83ac14ffbebc6c23ea744d0836a176405)

discourse/plugins: drop unused plugins

(cherry picked from commit 313f98e3b04a89fdd8ad2ec140e515f095437f69)

discourse.plugins.discourse-events: remove from all-plugins

Dropped in a5c9ab878f87d9ac1a03c7dd3fc7def1e1f329bb.

(cherry picked from commit aafb45f6de713b588e032cec779834c95f9eeede)

discourse: 3.1.0.beta4 -> 3.1.0.beta5

(cherry picked from commit bc3e2760c003a7d933133fa681b896432b845637)

discourse/plugins: update all

(cherry picked from commit e6d56dc744c507d36f43a5d8091f8a951aa3a741)

discourse: 3.1.0.beta5 -> 3.1.0

(cherry picked from commit 25f659d236d02efdc1cb6d7ddd47d18ac68fe4e2)

discourse: update plugins

(cherry picked from commit 1b625feda1589fa48df8048cc6739090f593c8e1)

discourse-prometheus: remove applied patch

(cherry picked from commit 73c6e1acaa7d984f4c9de23ac770a015d256fb42)

discourse: 3.2.0.beta1 -> 3.2.0.beta3

(cherry picked from commit 23cbecaba66cf3dea24f64e10e0985e5908c02f6)

discourse: fix merge error

discourse: fixes for 23.11

discourse: 3.2.0.beta3 -> 3.3.0.beta2

discourse/plugins: 3.2.0.beta3 -> 3.3.0.beta2

+++ pkgs/servers/web-apps/discourse/update.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env nix-shell
-#! nix-shell -i python3 -p "python3.withPackages (ps: with ps; [ requests click click-log packaging ])" bundix bundler nix-update nurl prefetch-yarn-deps
+#! nix-shell -i python3 -p "python3.withPackages (ps: with ps; [ requests click click-log packaging ])" bundix bundler nix-update nurl prefetch-yarn-deps yarn2nix
+#! nix-shell -i python3 -p python3 python3Packages.requests python3Packages.click python3Packages.click-log bundix bundler nix-update nix-universal-prefetch prefetch-yarn-deps yarn2nix
 from __future__ import annotations
 
 import click
@@ -59,16 +60,20 @@ class DiscourseVersion:
 
 
 class DiscourseRepo:
+    gh_token = None
     version_regex = re.compile(r'^v\d+\.\d+\.\d+(\.beta\d+)?$')
     _latest_commit_sha = None
 
     def __init__(self, owner: str = 'discourse', repo: str = 'discourse'):
         self.owner = owner
         self.repo = repo
+        self.headers = None
+        if self.gh_token is not None:
+            self.headers = {'Authorization': 'token %s' % self.gh_token}
 
     @property
     def versions(self) -> Iterable[str]:
-        r = requests.get(f'https://api.github.com/repos/{self.owner}/{self.repo}/git/refs/tags').json()
+        r = requests.get(f'https://api.github.com/repos/{self.owner}/{self.repo}/git/refs/tags', headers=self.headers).json()
         tags = [x['ref'].replace('refs/tags/', '') for x in r]
 
         # filter out versions not matching version_regex
@@ -80,7 +85,7 @@ class DiscourseRepo:
     @property
     def latest_commit_sha(self) -> str:
         if self._latest_commit_sha is None:
-            r = requests.get(f'https://api.github.com/repos/{self.owner}/{self.repo}/commits?per_page=1')
+            r = requests.get(f'https://api.github.com/repos/{self.owner}/{self.repo}/commits?per_page=1', headers=self.headers)
             r.raise_for_status()
             self._latest_commit_sha = r.json()[0]['sha']
 
@@ -100,7 +105,7 @@ class DiscourseRepo:
         :param str rev: the rev to fetch at :return:
 
         """
-        r = requests.get(f'https://raw.githubusercontent.com/{self.owner}/{self.repo}/{rev}/{filepath}')
+        r = requests.get(f'https://raw.githubusercontent.com/{self.owner}/{self.repo}/{rev}/{filepath}', headers=self.headers)
         r.raise_for_status()
         return r.text
 
@@ -234,21 +239,22 @@ def update(rev):
     _call_nix_update('discourse', version.version)
 
     old_yarn_hash = _nix_eval('discourse.assets.yarnOfflineCache.outputHash')
-    new_yarn_hash = repo.get_yarn_lock_hash(version.tag, "app/assets/javascripts/yarn-ember5.lock")
-    click.echo(f"Updating yarn lock hash: {old_yarn_hash} -> {new_yarn_hash}")
-
-    old_yarn_dev_hash = _nix_eval('discourse.assets.yarnDevOfflineCache.outputHash')
-    new_yarn_dev_hash = repo.get_yarn_lock_hash(version.tag, "yarn.lock")
-    click.echo(f"Updating yarn dev lock hash: {old_yarn_dev_hash} -> {new_yarn_dev_hash}")
-
+    new_yarn_hash = repo.get_yarn_lock_hash(version.tag, "yarn.lock")
+    click.echo(f"Updating yarn lock hash, {old_yarn_hash} -> {new_yarn_hash}")
     with open(Path(__file__).parent / "default.nix", 'r+') as f:
         content = f.read()
         content = content.replace(old_yarn_hash, new_yarn_hash)
-        content = content.replace(old_yarn_dev_hash, new_yarn_dev_hash)
         f.seek(0)
         f.write(content)
         f.truncate()
 
+    for fn in ['package.json', 'yarn.lock']:
+        with open(Path(__file__).parent / fn, 'w') as f:
+            f.write(repo.get_file(fn, version.tag))
+
+    yarnNix = subprocess.check_output(['yarn2nix'], text=True, cwd=Path(__file__).parent)
+    with open(Path(__file__).parent / 'yarn.nix', 'w') as f:
+        f.write(yarnNix)
 
 @cli.command()
 @click.argument('rev', default='latest')
@@ -273,25 +279,40 @@ def update_mail_receiver(rev):
 def update_plugins():
     """Update plugins to their latest revision."""
     plugins = [
+        {'name': 'discourse-adplugin'},
+        {'name': 'discourse-akismet'},
         {'name': 'discourse-assign'},
+        {'name': 'discourse-automation'},
         {'name': 'discourse-bbcode-color'},
+        {'name': 'discourse-cakeday'},
         {'name': 'discourse-calendar'},
         {'name': 'discourse-canned-replies'},
         {'name': 'discourse-chat-integration'},
         {'name': 'discourse-checklist'},
         {'name': 'discourse-data-explorer'},
         {'name': 'discourse-docs'},
+        {'name': 'discourse-follow'},
+        {'name': 'discourse-footnote'},
+        {'name': 'discourse-gamification'},
         {'name': 'discourse-github'},
+        {'name': 'discourse-graphviz'},
         {'name': 'discourse-ldap-auth', 'owner': 'jonmbake'},
+        {'name': 'discourse-linkedin-auth'},
         {'name': 'discourse-math'},
         {'name': 'discourse-migratepassword', 'owner': 'discoursehosting'},
         {'name': 'discourse-openid-connect'},
+        {'name': 'discourse-policy'},
         {'name': 'discourse-prometheus'},
         {'name': 'discourse-reactions'},
+        {'name': 'discourse-saml', 'keep_gemfile': False},
+        {'name': 'discourse-saved-searches'},
         {'name': 'discourse-saved-searches'},
         {'name': 'discourse-solved'},
         {'name': 'discourse-spoiler-alert'},
+        {'name': 'discourse-subscriptions'},
+        {'name': 'discourse-tooltips'},
         {'name': 'discourse-voting'},
+        {'name': 'discourse-whos-online'},
         {'name': 'discourse-yearly-review'},
     ]
 
@@ -300,6 +321,7 @@ def update_plugins():
         owner = plugin.get('owner') or "discourse"
         name = plugin.get('name')
         repo_name = plugin.get('repo_name') or name
+        keep_gemfile = plugin.get('keep_gemfile') or False
 
         if fetcher == "fetchFromGitHub":
             url = f"https://github.com/{owner}/{repo_name}"
@@ -419,14 +441,17 @@ def update_plugins():
                         f.write(content)
 
         if len(gemfile_text) > 0:
-            if os.path.isfile(gemfile):
+            if os.path.isfile(gemfile) and keep_gemfile is False:
                 os.remove(gemfile)
 
-            subprocess.check_output(['bundle', 'init'], cwd=rubyenv_dir)
+            if keep_gemfile is False:
+                subprocess.check_output(['bundle', 'init'], cwd=rubyenv_dir)
+
             os.chmod(gemfile, stat.S_IREAD | stat.S_IWRITE | stat.S_IRGRP | stat.S_IROTH)
 
             with open(gemfile, 'a') as f:
-                f.write(gemfile_text)
+                if keep_gemfile is False:
+                    f.write(gemfile_text)
 
             subprocess.check_output(['bundle', 'lock', '--add-platform', 'ruby'], cwd=rubyenv_dir)
             subprocess.check_output(['bundle', 'lock', '--update'], cwd=rubyenv_dir)
